refactor: guide 페이지 공통 컴포넌트로 분리 및 관련 상수 정리
