+++ src/app/ai/goal/guide/page.tsx
@@ -1,0 +1,8 @@
+import GuidePage from '@/src/app/ai/_components/GuidePage';
+import { ROUTINE_GUIDE_CONTENT } from '@/src/app/ai/_constants/guide';
+
+function RoutineGuidePage() {
+  return <GuidePage content={ROUTINE_GUIDE_CONTENT} />;
+}
+
+export default RoutineGuidePage;
